backend/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
requests>=2.31.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
whitenoise>=6.6.0
google-generativeai>=0.3.0
